requirements.txt
flask==2.3.3
wordfreq[cjk]==3.1.1
mecab-python3==1.0.9
